requirements.txt
daskms
numpy
matplotlib.pyplot
scipy.optimize
matplotlib.colors
astropy
PACK
dask.array
dask.distributed
statistics
statsmodels
scipy.signal
csv
tabulate
sys
random
mpl_toolkits.mplot3d
matplotlib.animation
datetime
